added clickoutside function to sidemenu and usermenu
