+++ package.json
@@ -12,6 +12,7 @@
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
     "react-loader-spinner": "^4.0.0",
+    "react-onclickoutside": "^6.12.0",
     "react-router-dom": "5.3.0",
     "react-scripts": "4.0.3",
     "react-transition-group": "^4.4.2",
